+++ src/app/modules/shared/components/cedar-embeddable-metadata-editor/cedar-embeddable-metadata-editor.component.ts
@@ -27,7 +27,7 @@ import { RenderSchedulerService } from '../../service/render-scheduler.service';
   standalone: false,
 })
 export class CedarEmbeddableMetadataEditorComponent implements OnDestroy {
-  private static INNER_VERSION = '2026-09-04 21:20 ed890758';
+  private static INNER_VERSION = '2026-09-06 09:16 f8ca7871';
 
   dataContext: DataContext | null = null;
   handlerContext: HandlerContext | null = null;